refactor: reorganize class
