+++ lib/core/design/widgets/auth_error_view.dart
