+++ lib/core/design/widgets/no_user_view.dart
