+++ lib/features/map/presentation/widgets/map_story_list_view.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:storyzz/core/data/networking/responses/list_story.dart';
+import 'package:storyzz/core/design/widgets/empty_story.dart';
+import 'package:storyzz/core/design/widgets/story_error_view.dart';
 import 'package:storyzz/core/providers/app_provider.dart';
 import 'package:storyzz/core/providers/auth_provider.dart';
 import 'package:storyzz/core/providers/story_provider.dart';
-import 'package:storyzz/core/widgets/empty_story.dart';
-import 'package:storyzz/core/widgets/story_error_view.dart';
 import 'package:storyzz/features/map/presentation/widgets/map_story_card.dart';
 import 'package:storyzz/features/map/provider/map_provider.dart';
